feat: Update Footer and Results components
The Footer component now has a minimum height of 10 pixels, and the "Docente" text has been removed. In the Results component, a new prop "power" has been added, and the descriptions for the "et" and "hps" inputs have been updated. Additionally, a new Environmental component has been added to the Content component.